+++ src/main/st/org/scaloid/common/package.scala
@@ -1,41 +1,10 @@
 package org.scaloid
 
 import android.app._
-import admin.DevicePolicyManager
-import android.view._
-import android.net._
-import android.os._
-import android.media._
-import collection.mutable.ArrayBuffer
-import android.util.Log
-
-import android.text._
-import android.view.accessibility._
-import android.accounts._
-import android.view.inputmethod._
-import android.location._
-import android.hardware._
-import android.telephony._
-import android.net.wifi._
 import android.content._
-import android.widget._
-import android.inputmethodservice._
-import android.preference._
-import android.preference.Preference._
-import android.view.WindowManager.LayoutParams._
-import android.view.View._
-import android.graphics.drawable.Drawable
-import java.lang.CharSequence
-import scala.Int
-import android.view.ContextMenu.ContextMenuInfo
-import android.text.method._
-import android.gesture._
-import android.appwidget._
+import android.os._
+import android.view._
 import annotation.target.{beanGetter, getter}
-import android.view.ViewGroup.LayoutParams
-import android.widget.TextView.OnEditorActionListener
-import android.graphics._
-import android.opengl._
 
 import language.implicitConversions
 
@@ -46,21 +15,7 @@ import language.implicitConversions
  *
  * @author Sung-Ho Lee
  */
-package object common extends Logger with SystemService with Implicits {
-
-  /**
-   * Launches a new activity for a give uri. For example, opens a web browser for http protocols.
-   */
-  def openUri(uri: Uri)(implicit context: Context) {
-    context.startActivity(new Intent(Intent.ACTION_VIEW, uri))
-  }
-
-  def play(uri: Uri = notificationSound)(implicit context: Context) {
-    val r = RingtoneManager.getRingtone(context, uri)
-    if (r != null) {
-      r.play()
-    }
-  }
+package object common extends Logger with SystemService with Helpers with Implicits {
 
   val idSequence = new java.util.concurrent.atomic.AtomicInteger(0)
 
@@ -71,155 +26,7 @@ package object common extends Logger with SystemService with Implicits {
     } while(activity.findViewById(candidate) != null)
     candidate
   }
-  
-  class UnitConversion(val ext: Double)(implicit context: Context) {
-    def dip: Int = (ext * context.getResources().getDisplayMetrics().density).toInt
-    def sp : Int = (ext * context.getResources().getDisplayMetrics().scaledDensity).toInt
-  }
 
-  @inline implicit def Double2unitConversion(ext: Double)(implicit context: Context): UnitConversion = new UnitConversion(ext)(context)
-  @inline implicit def Long2unitConversion  (ext: Long)  (implicit context: Context): UnitConversion = new UnitConversion(ext)(context)
-  @inline implicit def Int2unitConversion   (ext: Int)   (implicit context: Context): UnitConversion = new UnitConversion(ext)(context)
-
-  class ResourceConversion(val id: Int)(implicit context: Context) {
-    def r2Text         : CharSequence        = context.getText(id)
-    def r2TextArray    : Array[CharSequence] = context.getResources.getTextArray(id)
-    def r2String       : String              = context.getResources.getString(id)
-    def r2StringArray  : Array[String]       = context.getResources.getStringArray(id)
-    def r2Drawable     : Drawable            = context.getResources.getDrawable(id)
-    def r2Movie        : Movie               = context.getResources.getMovie(id)
-  }
-
-  @inline implicit def Int2resource(ext: Int)(implicit context: Context): ResourceConversion = new ResourceConversion(ext)(context)
-
-  // r2String is not provided because it is ambiguous with r2Text
-  @inline implicit def r2Text       (id: Int)(implicit context: Context): CharSequence        = context.getText(id)
-  @inline implicit def r2TextArray  (id: Int)(implicit context: Context): Array[CharSequence] = context.getResources.getTextArray(id)
-  @inline implicit def r2StringArray(id: Int)(implicit context: Context): Array[String]       = context.getResources.getStringArray(id)
-  @inline implicit def r2Drawable   (id: Int)(implicit context: Context): Drawable            = context.getResources.getDrawable(id)
-  @inline implicit def r2Movie      (id: Int)(implicit context: Context): Movie               = context.getResources.getMovie(id)
-
-  @inline implicit def string2Uri           (str: String): Uri            = Uri.parse(str)
-  @inline implicit def string2IntentFilter  (str: String): IntentFilter   = new IntentFilter(str)
-
-
-  implicit def func2runnable[F](f: () => F): Runnable =
-    new Runnable() {
-      def run() {
-        f()
-      }
-    }
-
-  implicit def lazy2runnable[F](f: => F): Runnable =
-    new Runnable() {
-      def run() {
-        f
-      }
-    }
-
-  $wholeClassDef(android.preference.Preference)$
-  $richClassDef(android.preference.DialogPreference)$
-  $wholeClassDef(android.preference.EditTextPreference)$
-
-  class AlertDialogBuilder(_title: CharSequence = null, _message: CharSequence = null)(implicit context: Context) extends AlertDialog.Builder(context) {
-    if (_title != null) setTitle(_title)
-    if (_message != null) setMessage(_message)
-
-
-    @inline def positiveButton(name: CharSequence = android.R.string.yes, onClick: => Unit = {}): AlertDialogBuilder =
-      positiveButton(name, (_, _) => {
-        onClick
-      })
-
-    @inline def positiveButton(name: CharSequence, onClick: (DialogInterface, Int) => Unit): AlertDialogBuilder = {
-      setPositiveButton(name, func2DialogOnClickListener(onClick))
-      this
-    }
-
-    @inline def neutralButton(name: CharSequence = android.R.string.ok, onClick: => Unit = {}): AlertDialogBuilder =
-      neutralButton(name, (_, _) => {
-        onClick
-      })
-
-    @inline def neutralButton(name: CharSequence, onClick: (DialogInterface, Int) => Unit): AlertDialogBuilder = {
-      setNeutralButton(name, func2DialogOnClickListener(onClick))
-      this
-    }
-
-    @inline def negativeButton(name: CharSequence, onClick: => Unit): AlertDialogBuilder =
-      negativeButton(name, (_, _) => {
-        onClick
-      })
-
-    @inline def negativeButton(name: CharSequence = android.R.string.no, onClick: (DialogInterface, Int) => Unit = (d, _) => {
-      d.cancel()
-    }): AlertDialogBuilder = {
-      setNegativeButton(name, func2DialogOnClickListener(onClick))
-      this
-    }
-
-    var tit: CharSequence = null
-
-    @inline def title_=(str: CharSequence) = {
-      tit = str
-      setTitle(str)
-    }
-
-    @inline def title = tit
-
-    var msg: CharSequence = null
-
-    @inline def message_=(str: CharSequence) = {
-      tit = str
-      setMessage(str)
-    }
-
-    @inline def message = tit
-
-    override def show():AlertDialog = runOnUiThread(super.show())
-  }
-
-  @inline def alert(title: CharSequence, text: CharSequence, clickCallback: => Unit = {})(implicit context: Context) {
-    new AlertDialogBuilder(title, text) {
-      neutralButton(android.R.string.ok, clickCallback)
-    }.show()
-  }
-
-  object SIntent {
-    @inline def apply[T]()(implicit context: Context, mt: ClassManifest[T]) = new Intent(context, mt.erasure)
-
-    @inline def apply[T](action: String)(implicit context: Context, mt: ClassManifest[T]): Intent = SIntent[T].setAction(action)
-  }
-
-
-  @inline def toast(message: CharSequence)(implicit context: Context) {
-    runOnUiThread(Toast.makeText(context, message, Toast.LENGTH_SHORT).show())
-  }
-
-  @inline def longToast(message: CharSequence)(implicit context: Context) {
-    runOnUiThread(Toast.makeText(context, message, Toast.LENGTH_LONG).show())
-  }
-
-  @inline def spinnerDialog(title: String, message: String)(implicit context: Context): ProgressDialog =
-    runOnUiThread(ProgressDialog.show(context, title, message, true))
-
-  @inline def pendingService(intent: Intent)(implicit context: Context) =
-    PendingIntent.getService(context, 0, intent, 0)
-
-  @inline def pendingActivity(intent: Intent)(implicit context: Context) =
-    PendingIntent.getActivity(context, 0, intent, 0)
-
-  @inline def pendingActivity[T](implicit context: Context, mt: ClassManifest[T]) =
-    PendingIntent.getActivity(context, 0, SIntent[T], 0)
-
-  @inline def notificationSound: Uri = RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION)
-
-  @inline def ringtoneSound: Uri = RingtoneManager.getDefaultUri(RingtoneManager.TYPE_RINGTONE)
-
-  @inline def alarmSound: Uri = RingtoneManager.getDefaultUri(RingtoneManager.TYPE_ALARM)
-
-  @inline def defaultSharedPreferences(implicit context: Context): SharedPreferences =
-    PreferenceManager.getDefaultSharedPreferences(context)
 
   /**
    * Provides handler instance and runOnUiThread() utility method.
@@ -241,14 +48,9 @@ package object common extends Logger with SystemService with Implicits {
     }
   }
 
-  def broadcastReceiver(filter: IntentFilter)(onReceiveBody: (Context, Intent) => Any)(implicit ctx: Context, reg: Registerable) {
-    val receiver = new BroadcastReceiver {
-      def onReceive(context: Context, intent: Intent) {
-        onReceiveBody(context, intent)
-      }
-    }
-    reg.onRegister(ctx.registerReceiver(receiver, filter))
-    reg.onUnregister(ctx.unregisterReceiver(receiver))
-  }
+  @getter
+  @beanGetter
+  private[scaloid] class noEquivalentGetterExists extends annotation.StaticAnnotation
+  private[scaloid] trait NoGetterForThisProperty
 }
